Move the node stuff to top level, now that we got rid of most of the old stuff.
